Adjustments to genres processing

+++ src/main/kotlin/io/github/luksal/book/db/document/book/BookStats.kt
@@ -1,6 +1,6 @@
 package io.github.luksal.book.db.document.book
 
 data class BookRatingCountStats(
-    val ratingCount: Int,
+    val ratingCount: Long,
     val documents: Long,
 )
