+++ src/main/kotlin/io/github/luksal/statistics/api/dto/StatisticsDto.kt
@@ -30,7 +30,7 @@ data class BookRatingStatisticsDto(
 )
 
 data class BookRatingValueDto(
-    val numberOfRatings: Int,
+    val numberOfRatings: Long,
     val value: Long,
 )
